Split up NextAuth config and install Payload plugin
