+++ next-frontend/src/auth.ts
@@ -1,40 +1,8 @@
-import NextAuth from "next-auth";
+import NextAuth from "next-auth"
+import { withPayload } from "payload-authjs";
+import payloadConfig from "@payload-config";
+import { authConfig } from "@/auth.config";
 
-export const { handlers, signIn, signOut, auth } = NextAuth({
-  debug: true,
-  providers: [
-    {
-      id: "WCA",
-      name: "WCA-OIDC-Provider",
-      type: "oidc",
-      issuer: process.env.OIDC_ISSUER,
-      clientId: process.env.OIDC_CLIENT_ID,
-      clientSecret: process.env.OIDC_CLIENT_SECRET,
-    },
-  ],
-  callbacks: {
-    async jwt({ token, account }) {
-      if (account) {
-        // First-time login, save the `access_token`, its expiry and the `refresh_token`
-        return {
-          ...token,
-          userId: account.userId,
-          access_token: account.access_token,
-          expires_at: account.expires_at,
-          refresh_token: account.refresh_token,
-        };
-      } else if (Date.now() < (token.expires_at as number) * 1000) {
-        // Subsequent logins, but the `access_token` is still valid
-        return token;
-      } else {
-        // TODO Implement Refreshing
-        return token;
-      }
-    },
-    async session({ session, token }) {
-      session.accessToken = token.access_token;
-      session.user.id = token.userId;
-      return session;
-    },
-  },
-});
+export const { handlers, signIn, signOut, auth } = NextAuth(
+  withPayload(authConfig, { payloadConfig })
+)
